requirements.txt
fastapi
uvicorn
# Install CPU versions from: https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu
torchvision==0.19.1+cpu
transformers==4.46.3
datasets
ipfshttpclient==0.8.0a2
cryptography==42.0.8
flwr==1.10.0
librosa
pyttsx3
numpy
pillow
requests
python-dotenv
pydub
pytest
huggingface_hub[hf_xet]  # Optional for faster Hugging Face downloads
python-multipart
